requirements-graphql.txt
# GraphQL API Requirements
graphene-django>=3.2.0
graphene-django-optimizer>=0.10.0
graphene-file-upload>=1.3.0
django-graphql-jwt>=0.4.0
graphql-core>=3.2.0
django-cors-headers>=4.3.0
